stockbackend/requirements.txt
Django==4.2
djangorestframework==3.14.0
pandas==2.2.1
yfinance==0.2.36
pytz==2024.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.8.2
requests